refactor: utils 분리
